Added plugin class; removed plugin code from PS2Addon
